\name{mafun-package}
\alias{mafun-package}
\alias{mafun}
\docType{package}
\title{
Collection of functions used in Master's Thesis
}
\description{
It provides a clean overview over the functions I wrote during my Master's Thesis. It is mainly concentrated on the work with Twitter Data in R.
}
\details{
\tabular{ll}{
Package: \tab mafun\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-15\cr
License: \tab MIT License\cr
}
See Index \href{http://127.0.0.1:24911/help/library/mafun/html/00Index.html}{Index} for an overview over the included functions.
}
\author{
Soeren Abbenseth | spam at soeren . me