% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EmoticonScore}
\alias{EmoticonScore}
\title{EmoticonScore}
\usage{
EmoticonScore(doc, smile = smile, frown = frown)
}
\arguments{
\item{doc}{A character string OR a vector of character strings.}

\item{smile}{A vector of smile-emoticons.}

\item{frown}{A vector of frown-emoticons.}
}
\description{
This function classifies a document as "positive", "negative" or "neutral" on the occurances of smile and frown emoticons.
}
\details{
The functions searches through the text of given documents and returns +1 if the text contains a smile-emoticon,
-1 if it contains a frown emoticon and 0 if no emoticons were found in the text.

The functions provides default vectors for smile- and frown-emoticons, inspired by the Read (2005), which can be overwritten with self-defined vectors.
To see the vectors, type "EmoticonScore" (without brackets) into the console and press Enter.
}
\examples{
tweet1 <- 'I am totally happy :-)'
tweet2 <- 'I feel neutral'
tweet3 <- 'I am totally sad =('

tweets <- c(tweet1, tweet2, tweet3)

sapply(tweets, EmoticonScore)
I am totally happy :-)         I feel neutral    I am totally sad =(
                     1                      0                     -1
}
\seealso{
Read, Jonathon, ed. (2005): Using Emoticons to reduce Dependency in Machine Learning Techniques for Sentiment Classification. ACL Student Research Workshop. (2005). Association for Computational Linguistics. Ann Arbor.
}

