% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTweets}
\alias{getTweets}
\title{getTweets}
\usage{
getTweets(dir = getwd(), varsToKeep = basicVars,
  restrictToPartyTweets = TRUE, restrictToLang = "de", wordlistPositive,
  wordlistNegative, combine = TRUE)
}
\arguments{
\item{bla}{}
}
\description{
bla
}
\details{
bla
}
\examples{
bla
}

